Refresh contents of the looked at container once a second
Previously the contents would be fetched only when the player starts staring at the container. But the contents would not get refreshed if someone edits the container contents. Also, containers like furnaces and hoppers update their contents actively.

Ideally I would just listen to appropriate update events from those containers, but it looks like that is going to get complicated because there are so many different ways that a container can update.